Split initializers away from call to test ENV
